backend/requirements.txt
flask==2.2.5
flask-cors==4.0.0
flask-jwt-extended==4.5.2
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
passlib==1.7.4
networkx==3.1
matplotlib==3.7.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
supabase==2.3.0
gunicorn==21.2.0
pytest==7.4.3
python-json-logger==2.0.7
email-validator==2.1.0
Werkzeug==2.2.3
requests==2.31.0
bcrypt==4.0.1 